+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Started Change log file as per release 0.0.0.54
 
+## AGMPowerLIB  (0.0.0.59)
+* Get-AGMLibPolicies - added options 
+
 ## AGMPowerLIB  (0.0.0.57)
 * New-AGMLibGCPInstance 
   * was printing a dummy instancename for multi-instance recovery, this was causing issue, so field will be blank